refactor: extract button helpers
